+++ examples/wakustealthcommitments/node_spec.nim
@@ -1,7 +1,4 @@
-when (NimMajor, NimMinor) < (1, 4):
-  {.push raises: [Defect].}
-else:
-  {.push raises: [].}
+{.push raises: [].}
 
 import
   ../../waku/common/logging, ../../waku/factory/[waku, networks_config, external_config]