+++ tests/waku_rln_relay/test_rln_group_manager_static.nim
@@ -1,9 +1,6 @@
 {.used.}
 
-when (NimMajor, NimMinor) < (1, 4):
-  {.push raises: [Defect].}
-else:
-  {.push raises: [].}
+{.push raises: [].}
 
 import
   testutils/unittests,