+++ waku/common/confutils/envvar/defs.nim
@@ -1,7 +1,4 @@
-when (NimMajor, NimMinor) < (1, 4):
-  {.push raises: [Defect].}
-else:
-  {.push raises: [].}
+{.push raises: [].}
 
 import confutils/defs as confutilsDefs
 import ../../envvar_serialization