+++ waku/common/logging.nim
@@ -8,10 +8,7 @@ import
 
 export chronicles.LogLevel
 
-when (NimMajor, NimMinor) < (1, 4):
-  {.push raises: [Defect].}
-else:
-  {.push raises: [].}
+{.push raises: [].}
 
 type LogFormat* = enum
   TEXT