+++ waku/common/utils/matterbridge_client.nim
@@ -1,7 +1,4 @@
-when (NimMajor, NimMinor) < (1, 4):
-  {.push raises: [Defect].}
-else:
-  {.push raises: [].}
+{.push raises: [].}
 
 import std/[httpclient, json, uri, options], stew/results