+++ waku/factory/node_factory.nim
@@ -135,17 +135,21 @@ proc setupProtocols(
         # only peers with populated records
         .mapIt(toRemotePeerInfo(it.record.get()))
 
-      debug "connecting to exchanged peers",
+      debug "adding exchanged peers",
         src = peer, topic = topic, numPeers = exchangedPeers.len
 
-      # asyncSpawn, as we don't want to block here
-      asyncSpawn node.connectToNodes(exchangedPeers, "peer exchange")
+      for peer in exchangedPeers:
+        # Peers added are filtered by the peer manager
+        node.peerManager.addPeer(peer, PeerOrigin.PeerExchange)
 
     peerExchangeHandler = some(handlePeerExchange)
 
+  let shards =
+    conf.contentTopics.mapIt(node.wakuSharding.getShard(it).expect("Valid Shard"))
+  debug "Shards created from content topics",
+    contentTopics = conf.contentTopics, shards = shards
+
   if conf.relay:
-    let shards =
-      conf.contentTopics.mapIt(node.wakuSharding.getShard(it).expect("Valid Shard"))
     let pubsubTopics = conf.pubsubTopics & shards
 
     let parsedMaxMsgSize = parseMsgSize(conf.maxMessageSize).valueOr: