+++ waku/node/waku_node.nim
@@ -1,7 +1,4 @@
-when (NimMajor, NimMinor) < (1, 4):
-  {.push raises: [Defect].}
-else:
-  {.push raises: [].}
+{.push raises: [].}
 
 import
   std/[hashes, options, sugar, tables, strutils, sequtils, os, net],
@@ -437,12 +434,14 @@ proc mountFilter*(
       filter_subscriptions.DefaultSubscriptionTimeToLiveSec,
     maxFilterPeers: uint32 = filter_subscriptions.MaxFilterPeers,
     maxFilterCriteriaPerPeer: uint32 = filter_subscriptions.MaxFilterCriteriaPerPeer,
+    messageCacheTTL: Duration = filter_subscriptions.MessageCacheTTL,
 ) {.async: (raises: []).} =
   ## Mounting filter v2 protocol
 
   info "mounting filter protocol"
   node.wakuFilter = WakuFilter.new(
-    node.peerManager, subscriptionTimeout, maxFilterPeers, maxFilterCriteriaPerPeer
+    node.peerManager, subscriptionTimeout, maxFilterPeers, maxFilterCriteriaPerPeer,
+    messageCacheTTL,
   )
 
   if node.started: