+++ waku/waku_archive/archive.nim
@@ -1,7 +1,4 @@
-when (NimMajor, NimMinor) < (1, 4):
-  {.push raises: [Defect].}
-else:
-  {.push raises: [].}
+{.push raises: [].}
 
 import
   std/[times, options, sequtils, strutils, algorithm],
@@ -141,7 +138,7 @@ proc findMessages*(
 
   let isAscendingOrder = query.direction.into()
 
-  if query.contentTopics.len > 10:
+  if query.contentTopics.len > 100:
     return err(ArchiveError.invalidQuery("too many content topics"))
 
   if query.cursor.isSome() and query.cursor.get().hash.len != 32:
@@ -229,7 +226,7 @@ proc findMessagesV2*(
 
   let isAscendingOrder = query.direction.into()
 
-  if query.contentTopics.len > 10:
+  if query.contentTopics.len > 100:
     return err(ArchiveError.invalidQuery("too many content topics"))
 
   let queryStartTime = getTime().toUnixFloat()