+++ waku/waku_archive/driver/queue_driver/index.nim
@@ -1,7 +1,4 @@
-when (NimMajor, NimMinor) < (1, 4):
-  {.push raises: [Defect].}
-else:
-  {.push raises: [].}
+{.push raises: [].}
 
 import stew/byteutils, nimcrypto/sha2
 import ../../../waku_core, ../../common