+++ waku/waku_noise/noise_utils.nim
@@ -3,10 +3,7 @@
 ## See spec for more details:
 ## https://github.com/vacp2p/rfc/tree/master/content/docs/rfcs/35
 
-when (NimMajor, NimMinor) < (1, 4):
-  {.push raises: [Defect].}
-else:
-  {.push raises: [].}
+{.push raises: [].}
 
 import std/[algorithm, base64, oids, options, strutils, tables, sequtils]
 import chronos