+++ waku/waku_rln_relay/rln_relay.nim
@@ -1,7 +1,5 @@
-when (NimMajor, NimMinor) < (1, 4):
-  {.push raises: [Defect].}
-else:
-  {.push raises: [].}
+
+{.push raises: [].}
 
 import
   std/[sequtils, tables, times, deques],